% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{display_image}
\alias{display_image}
\title{Display color image}
\usage{
display_image(image)
}
\arguments{
\item{image}{Matrix The image from which the palette will be extracted from. Should
be a 3 (or more) dimensional matrix. The output of functions such as \code{readJPG()}
are suitable as \code{image}.}
}
\value{
A raster image in the plot window.
}
\description{
Convenience wrapper to create a raster image of the image
you wish to extract the palette from.
}
\examples{
\dontrun{
library(jpeg)
your_image <- readJPEG("path/to/your/image.jpg")
display_image(your_image)
}
}

