% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{display_palette}
\alias{display_palette}
\title{Display color palette}
\usage{
display_palette(palette)
}
\arguments{
\item{palette}{Vector The output of \code{image_palette}.}
}
\value{
A plot of the colors extracted from the image
}
\description{
Displays the created palette as a barchart with axis labels
representing hex values of the colors. A more attractive method
for doing so would be to use \code{show_cols()} from
\code{library(scales)}.
}
\examples{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
display_palette(image_palette(img, n=5))
}
\seealso{
\code{scales::show_cols()}
}

