% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale_image.R
\name{image_pal}
\alias{image_pal}
\title{Image palette}
\usage{
image_pal(image, choice = mean, volume = FALSE)
}
\arguments{
\item{image}{Matrix The image from which the palette will be extracted from. Should
be a 3 (or more) dimensional matrix. The output of a function such as \code{readJPG()}
or \code{readPNG()} are suitable as \code{image}.}

\item{choice}{Function Defines how the color will be chosen from the final color cubes.
The default choice is to take the \code{mean} value of the image cube, but other choices
may return a subjectively superior scale. Try \code{median}, or \code{min}, or \code{max}, or
whatever summary statistic suits your fancy.}

\item{volume}{Logical volume controls the method for choosing which color cube to split
at each iteration of the algorithm. The default choice (when \code{volume = FALSE}) is to
choose the cube based on which cube contains the largest extent (that is, the largest range
of some color). When \code{volume = TRUE}, the cube with the largest volume is chosen to split.
Occasionally, setting to \code{TRUE} returns a better palette.}
}
\description{
Image palette function
}
\examples{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
display_image(img)
scales::show_col(image_pal(img)(10))
}

