% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/volumeBox.R
\name{vbox}
\alias{vbox}
\title{Volume box}
\usage{
vbox(im)
}
\arguments{
\item{im}{List An image in list form, with three components: red, green, blue}
}
\value{
A list containing the minimum, maximum, median, extent, and volume
 of each component of the image
}
\description{
Extract minimum, maximum, median, extent, and volume information
from red, green, and blue color channels.
}
\details{
For passing to \code{median_cut()}.
}
\seealso{
\code{\link[median_cut]{median_cut}} \code{\link[create_palette]{create_palette}}
}

