% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{display_palette}
\alias{display_palette}
\title{Display color palette}
\usage{
display_palette(palette)
}
\arguments{
\item{palette}{Vector The output of \code{create_palette}.}
}
\value{
A plot of the colors extracted from the image
}
\description{
Displays the created palette as a barchart with axis labels
representing hex values of the colors. A more attractive method
for doing so would be to use \code{show_cols()} from
\code{library(scales)}.
}
\examples{
\dontrun{
library(jpeg)
your_image <- readJPEG("path/to/your/image.jpg")
display_palette(create_palette(your_image, n=5))
}
}
\seealso{
\code{scales::show_cols()}
}

