% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale_image.R
\name{scale_color_image}
\alias{scale_color_image}
\alias{scale_colour_image}
\alias{scale_fill_image}
\title{Image color scales}
\usage{
scale_color_image(..., image, n = 5, choice = mean, volume = FALSE,
  discrete = TRUE)

scale_colour_image(..., image, n = 5, choice = mean, volume = FALSE,
  discrete = TRUE)

scale_fill_image(..., image, n = 5, choice = mean, volume = FALSE,
  discrete = TRUE)
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{image}{pass through parameter to \code{create_palette}}

\item{n}{For continuous color scales, you may optionally pass in an integer, n.
This allows some control over the scale, if n is too large the scale has too many
colors and ceases to be meaningul. n = 3 to n = 5 is recommended.}

\item{choice}{pass through parameter to \code{create_palette}}

\item{volume}{pass through parameter to \code{create_palette}}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}
}
\description{
Uses the image color scale.
}
\details{
For \code{discrete == TRUE} (the default) the function will return a \code{discrete_scale} with the plot-computed
number of colors. All other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.

See \link{create_palette} for more information on the color scale.
}
\examples{
\dontrun{
library(ggplot2)

# ripped from the pages of ggplot2
your_image <- readJPEG("path/to/your/image.jpg")
display_image(your_image)
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point(size=4, aes(colour = factor(cyl))) +
    scale_color_image(image = your_image) +
    theme_bw()

# ripped from the pages of ggplot2
dsub <- subset(diamonds, x > 5 & x < 6 & y > 5 & y < 6)
dsub$diff <- with(dsub, sqrt(abs(x-y))* sign(x-y))
d <- ggplot(dsub, aes(x, y, colour=diff)) + geom_point()
d + scale_color_image(image = your_image, discrete=FALSE) + theme_bw()


}
}
\seealso{
\code{\link{median_cut}} \code{\link{create_palette}} \code{\link{vbox}}
\code{\link{display_image}}
}

