% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_colors}
\alias{switch_colors}
\title{Swap Colors in an Image}
\usage{
switch_colors(target_image, source_image, source_colors = 3,
  smoothness = 100, ...)
}
\arguments{
\item{target_image}{Matrix The image you wish to transfer colors into.
The output from \code{readJPEG} is of suitable format.}

\item{source_image}{Matrix The image you wish to transfer colors from.}

\item{source_colors}{Integer The number of colors you wish to extract from the
source image.}

\item{smoothness}{Integer The source colors get interpolated so that the image
doesn't appear blocky. The higher the value, the smoother the output.}

\item{...}{Pass any of the arguments for \code{image_palette}}
}
\value{
The image, but with swapped colors!
}
\description{
Swap the palette of an image!
}
\examples{
\dontrun{
library(jpeg)
america <- jpeg::readJPEG("path/to/flagImage/AmericanFlag.jpg")
obama <- jpeg::readJPEG("path/to/obamaImage/Obama.jpg")
switch_colors(obama, america)
}
}

