% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_colors}
\alias{switch_colors}
\title{Swap Colors in an Image}
\usage{
switch_colors(im, palette)
}
\arguments{
\item{im}{Matrix An image matrix, such as the output from \code{readJPEG}
that you wish to swap a different palette into.}

\item{palette}{Vector The output of \code{create_palette}.}
}
\value{
The image, but with swapped colors!
}
\description{
Swap the palette of an image!
}
\examples{
\dontrun{
library(jpeg)
america <- jpeg::readJPEG("path/to/flagImage/AmericanFlag.jpg")
obama <- jpeg::readJPEG("path/to/obamaImage/Obama.jpg")
switch_colors(obama, sample(create_palette(america, choice=median, 4)))
}
}

